refactor copy to study actions, separate assay specific code
