+++ study/src/org/labkey/study/publish/AbstractPublishConfirmAction.java
@@ -1,0 +1,427 @@
+/*
+ * Copyright (c) 2008-2019 LabKey Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.labkey.study.publish;
+
+import org.apache.commons.beanutils.ConversionException;
+import org.apache.commons.beanutils.ConvertUtils;
+import org.jetbrains.annotations.Nullable;
+import org.labkey.api.action.FormViewAction;
+import org.labkey.api.action.SpringActionController;
+import org.labkey.api.data.ActionButton;
+import org.labkey.api.data.Container;
+import org.labkey.api.data.ContainerManager;
+import org.labkey.api.data.DataRegionSelection;
+import org.labkey.api.query.FieldKey;
+import org.labkey.api.query.QuerySettings;
+import org.labkey.api.query.UserSchema;
+import org.labkey.api.security.RequiresPermission;
+import org.labkey.api.security.permissions.InsertPermission;
+import org.labkey.api.study.Study;
+import org.labkey.api.study.StudyService;
+import org.labkey.api.study.StudyUrls;
+import org.labkey.api.study.TimepointType;
+import org.labkey.api.study.publish.PublishKey;
+import org.labkey.api.study.assay.AssayPublishService;
+import org.labkey.api.study.query.PublishResultsQueryView;
+import org.labkey.api.util.PageFlowUtil;
+import org.labkey.api.util.URLHelper;
+import org.labkey.api.view.ActionURL;
+import org.labkey.api.view.JspView;
+import org.labkey.api.view.RedirectException;
+import org.labkey.api.view.VBox;
+import org.labkey.api.view.template.ClientDependency;
+import org.springframework.validation.BindException;
+import org.springframework.validation.Errors;
+import org.springframework.web.servlet.ModelAndView;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+
+/**
+ * User: brittp
+ * Date: Jul 27, 2007
+ * Time: 11:02:04 AM
+ */
+@RequiresPermission(InsertPermission.class)
+public abstract class AbstractPublishConfirmAction<FORM extends PublishConfirmForm> extends FormViewAction<FORM>
+{
+    @Nullable protected String _targetStudyName;
+    protected Map<Object, String> _postedVisits;
+    protected Map<Object, String> _postedDates;
+    protected Map<Object, String> _postedPtids;
+    protected Map<Object, String> _postedTargetStudies;
+    protected Set<Integer> _selectedObjects;
+    protected List<Integer> _allObjects = Collections.emptyList();
+    protected Container _targetStudy;
+    protected ActionURL _successURL;
+
+    public static class PublishConfirmBean
+    {
+        private TimepointType _timepointType;
+        private final boolean _mismatched;
+
+        public PublishConfirmBean(TimepointType timepointType, boolean mismatched)
+        {
+            _timepointType = timepointType;
+            _mismatched = mismatched;
+        }
+
+        public TimepointType getTimepointType()
+        {
+            return _timepointType;
+        }
+
+        public boolean isMismatched()
+        {
+            return _mismatched;
+        }
+    }
+
+    @Override
+    public void validateCommand(FORM form, Errors errors)
+    {
+        // Check if a single target study was posted for the entire run (the common case)
+        if (form.getTargetStudy() != null && form.getTargetStudy().length == 1)
+        {
+            _targetStudy = ContainerManager.getForId(form.getTargetStudy()[0]);
+            if (_targetStudy == null)
+            {
+                errors.reject(SpringActionController.ERROR_MSG, "Could not find target study");
+            }
+        }
+
+        if (_targetStudy != null)
+        {
+            Study study = StudyService.get().getStudy(_targetStudy);
+            if (study == null)
+            {
+                errors.reject(SpringActionController.ERROR_MSG, "No study configured for " + _targetStudy);
+            }
+            AssayPublishService.get().getTimepointType(_targetStudy);
+            _targetStudyName = study.getLabel();
+        }
+
+        _selectedObjects = new HashSet<>(AbstractPublishStartAction.getCheckboxIds(getViewContext()));
+        _allObjects = form.getObjectIdValues();
+
+        if (_allObjects == null) // On first post, this is empty, so use the current selection
+            _allObjects = new ArrayList<>(_selectedObjects);
+    }
+
+    @Override
+    public boolean handlePost(FORM form, BindException errors) throws Exception
+    {
+        if (form.isAttemptPublish() && form.getDefaultValueSourceEnum() == PublishResultsQueryView.DefaultValueSource.UserSpecified)
+        {
+            _postedVisits = new HashMap<>();
+            _postedDates = new HashMap<>();
+            _postedPtids = new HashMap<>();
+            _postedTargetStudies = new HashMap<>();
+
+            attemptCopy(form, errors, _selectedObjects, _allObjects, _targetStudy, _postedTargetStudies, _postedVisits, _postedDates, _postedPtids);
+        }
+        return !errors.hasErrors();
+    }
+
+    @Override
+    public URLHelper getSuccessURL(FORM form)
+    {
+        return _successURL;
+    }
+
+    /**
+     * Create the schema for the underlying results table to display in the publish confirm view.
+     */
+    protected abstract UserSchema getUserSchema(FORM form);
+
+    /**
+     * Get the query settings to construct the publish confirm view
+     */
+    protected abstract QuerySettings getQuerySettings(FORM form);
+
+    /**
+     * Checks if the specimen participant/visit/dates from the source don't match based on the specimen id and target study
+     */
+    protected abstract boolean isMismatchedSpecimenInfo(FORM form);
+
+    /**
+     * Returns the publish URL
+     */
+    protected abstract ActionURL getPublishHandlerURL(FORM form);
+
+    /**
+     * The field key for the results objectId
+     */
+    protected abstract FieldKey getObjectIdFieldKey(FORM form);
+
+    /**
+     * Generate the map of field keys which will be added to publish results query view to represent the subject,
+     * timepoint editable columns etc.
+     */
+    protected abstract Map<PublishResultsQueryView.ExtraColFieldKeys, FieldKey> getAdditionalColumns(FORM form);
+
+    /**
+     * Returns the hidden form fields that need to be included on the data region form
+     */
+    protected abstract Map<String, Object> getHiddenFormFields(FORM form);
+
+    /**
+     * Perform the copy to study operation
+     */
+    protected abstract ActionURL copyToStudy(FORM form, Container targetStudy, Map<Integer, PublishKey> publishData, List<String> publishErrors);
+
+    @Override
+    public ModelAndView getView(FORM form, boolean reshow, BindException errors) throws Exception
+    {
+        getPageConfig().addClientDependency(ClientDependency.fromPath("study/assayPublish.js"));
+
+        UserSchema schema = getUserSchema(form);
+        boolean mismatched = isMismatchedSpecimenInfo(form);
+
+        // Show the form
+        QuerySettings settings = getQuerySettings(form);
+        settings.setAllowChooseView(false);
+        settings.setSelectionKey(form.getDataRegionSelectionKey());
+        if (form.getContainerFilterName() != null)
+            settings.setContainerFilterName(form.getContainerFilterName());
+        PublishResultsQueryView queryView = new PublishResultsQueryView(schema, settings, errors, getObjectIdFieldKey(form),
+                _allObjects, _targetStudy, _postedTargetStudies, _postedVisits, _postedDates, _postedPtids,
+                mismatched,
+                form.isIncludeTimestamp(),
+                getAdditionalColumns(form),
+                getHiddenFormFields(form));
+
+        List<ActionButton> buttons = new ArrayList<>();
+        URLHelper returnURL = form.getReturnURLHelper();
+        if (null == returnURL)
+        {
+//            returnURL = PageFlowUtil.urlProvider(AssayUrls.class).getAssayRunsURL(getContainer(), _protocol).addParameter("clearDataRegionSelectionKey", publishConfirmForm.getDataRegionSelectionKey());
+        }
+
+        ActionURL publishURL = getPublishHandlerURL(form);
+
+        publishURL.replaceParameter("defaultValueSource", PublishResultsQueryView.DefaultValueSource.UserSpecified.toString());
+        publishURL.replaceParameter("validate", "false");
+        ActionButton publishButton = new ActionButton(publishURL, "Copy to Study");
+        publishButton.setScript("return assayPublish_onCopyToStudy(this)", true);
+        buttons.add(publishButton);
+
+        publishURL.replaceParameter("validate", "true");
+        ActionButton validateButton = new ActionButton(publishURL, "Re-Validate");
+        validateButton.setScript("return assayPublish_onCopyToStudy(this)", true);
+        buttons.add(validateButton);
+
+        TimepointType timepointType = null;
+        if (_targetStudy != null)
+            timepointType = AssayPublishService.get().getTimepointType(_targetStudy);
+
+        if (timepointType != null && !timepointType.equals(TimepointType.VISIT))
+        {
+            publishURL.replaceParameter("defaultValueSource", PublishResultsQueryView.DefaultValueSource.Assay.toString());
+            publishURL.replaceParameter("includeTimestamp", "true");
+            ActionButton includeTimeButton = new ActionButton(publishURL, "Display DateTime");
+            includeTimeButton.setScript("return assayPublish_onCopyToStudy(this)", true);
+            buttons.add(includeTimeButton);
+        }
+
+        if (mismatched)
+        {
+            publishURL.deleteParameter("validate");
+            publishURL.replaceParameter("defaultValueSource", PublishResultsQueryView.DefaultValueSource.Assay.toString());
+            ActionButton fromAssayButton = new ActionButton(publishURL, "Reset with Assay Data");
+            buttons.add(fromAssayButton);
+
+            publishURL.replaceParameter("defaultValueSource", PublishResultsQueryView.DefaultValueSource.Specimen.toString());
+            ActionButton fromSpecimenButton = new ActionButton(publishURL, "Reset with Specimen Data");
+            buttons.add(fromSpecimenButton);
+        }
+
+        ActionButton cancelButton = new ActionButton("Cancel", returnURL);
+        cancelButton.setScript("LABKEY.setSubmit(true);", true);
+        buttons.add(cancelButton);
+
+        queryView.setButtons(buttons);
+
+        return new VBox(new JspView<>("/org/labkey/study/publish/view/publishHeader.jsp",
+                new PublishConfirmBean(timepointType, mismatched), errors), queryView);
+    }
+
+    private void attemptCopy(FORM form, BindException errors,
+                             Set<Integer> selectedObjects, List<Integer> allObjects,
+                             Container targetStudy,
+                             Map<Object, String> postedTargetStudies,
+                             Map<Object, String> postedVisits,
+                             Map<Object, String> postedDates,
+                             Map<Object, String> postedPtids)
+            throws RedirectException
+    {
+        Map<Integer, PublishKey> publishData = new LinkedHashMap<>();
+        String[] participantIds = form.getParticipantId();
+        String[] visitIds = form.getVisitId();
+        String[] dates = form.getDate();
+        String[] targetStudies = form.getTargetStudy();
+
+        Map<Object, Container> resolvedStudies = new HashMap<>();
+
+        boolean missingPtid = false;
+        boolean missingVisitId = false;
+        boolean missingDate = false;
+        boolean missingStudy = false;
+        boolean badVisitIds = false;
+        boolean badDates = false;
+        int index = 0;
+        for (int objectId : allObjects)
+        {
+            // we only want to give errors for selected rows, but we want to compute visits and ptids regardless
+            boolean selected = selectedObjects.contains(objectId);
+
+            String participantId = participantIds != null && participantIds.length > index ? participantIds[index] : null;
+            if (participantId == null || participantId.trim().length() == 0)
+            {
+                if (selected)
+                {
+                    missingPtid = true;
+                }
+            }
+            else
+                participantId = participantId.trim();
+
+            Container rowLevelTargetStudy = null;
+            if (targetStudies != null && targetStudies.length > index && null != targetStudies[index])
+            {
+                if (resolvedStudies.containsKey(targetStudies[index]))
+                    rowLevelTargetStudy = resolvedStudies.get(targetStudies[index]);
+                else
+                {
+                    Set<Study> studies = StudyService.get().findStudy(targetStudies[index], getUser());
+                    if (studies != null && !studies.isEmpty())
+                    {
+                        rowLevelTargetStudy = studies.iterator().next().getContainer();
+                        resolvedStudies.put(targetStudies[index], rowLevelTargetStudy);
+                    }
+                }
+            }
+
+            // No row level targetStudy found, use run level targetStudy.
+            if (rowLevelTargetStudy == null)
+                rowLevelTargetStudy = targetStudy;
+
+            if (rowLevelTargetStudy == null)
+            {
+                if (selected)
+                    missingStudy = true;
+            }
+            else
+            {
+                postedTargetStudies.put(objectId, rowLevelTargetStudy.getId());
+
+                if (AssayPublishService.get().getTimepointType(rowLevelTargetStudy) == TimepointType.VISIT)
+                {
+                    String visitIdStr = visitIds != null && visitIds.length > index ? visitIds[index] : null;
+                    Float visitId = null;
+                    if (visitIdStr == null || visitIdStr.trim().length() == 0)
+                    {
+                        if (selected)
+                            missingVisitId = true;
+                    }
+                    else
+                    {
+                        visitIdStr = visitIdStr.trim();
+                        try
+                        {
+                            visitId = Float.parseFloat(visitIdStr);
+                        }
+                        catch (NumberFormatException e)
+                        {
+                            if (selected)
+                                badVisitIds = true;
+                        }
+                    }
+                    postedPtids.put(objectId, participantId);
+                    postedVisits.put(objectId, visitIdStr);
+                    if (visitId != null && selected)
+                        publishData.put(objectId, new PublishKey(rowLevelTargetStudy, participantId, visitId.floatValue(), objectId));
+                }
+                else // TimepointType.DATE or CONTINUOUS
+                {
+                    String dateStr = dates != null && dates.length > index ? dates[index] : null;
+                    Date date = null;
+                    if (dateStr == null || dateStr.trim().length() == 0)
+                    {
+                        if (selected)
+                            missingDate = true;
+                    }
+                    else
+                    {
+                        dateStr = dateStr.trim();
+                        try
+                        {
+                            date = (Date) ConvertUtils.convert(dateStr, Date.class);
+                        }
+                        catch (ConversionException e)
+                        {
+                            if (selected)
+                                badDates = true;
+                        }
+                    }
+                    postedPtids.put(objectId, participantId);
+                    postedDates.put(objectId, dateStr);
+                    if (date != null && selected)
+                        publishData.put(objectId, new PublishKey(rowLevelTargetStudy, participantId, date, objectId));
+                }
+            }
+            index++;
+        }
+
+        if (missingStudy)
+            errors.reject(null, "You must specify a Target Study for all selected rows.");
+        if (missingPtid)
+            errors.reject(null, "You must specify a Participant ID for all selected rows.");
+        if (missingVisitId)
+            errors.reject(null, "You must specify a Visit ID for all selected rows.");
+        if (badVisitIds)
+            errors.reject(null, "Visit IDs must be numbers.");
+        if (missingDate || badDates)
+            errors.reject(null, "You must specify a Date for all selected rows.");
+        if (publishData.isEmpty())
+            errors.reject(null, "No data selected to publish");
+
+        if (errors.getErrorCount() == 0 && !form.isValidate())
+        {
+            List<String> publishErrors = new ArrayList<>();
+            _successURL  = copyToStudy(form, targetStudy, publishData, publishErrors);
+            if (publishErrors.isEmpty())
+            {
+                DataRegionSelection.clearAll(getViewContext(), form .getDataRegionSelectionKey());
+                // Issue 14111: successURL shouldn't be null if there are no errors, but better to go somewhere than to NPE.
+                if (_successURL == null)
+                    _successURL = PageFlowUtil.urlProvider(StudyUrls.class).getDatasetsURL(targetStudy);
+            }
+            for (String publishError : publishErrors)
+            {
+                errors.reject(null, publishError);
+            }
+        }
+    }
+}
